Restructure tests folder structure
