+++ tests/broadcasting/broadcastReblog.spec.ts
@@ -1,19 +1,20 @@
-import { broadcastFollow } from '../src/broadcastFollow';
-import * as broadcaster from '../src/broadcastOperations';
+import { broadcastReblog } from '../../src/broadcasting/broadcastReblog';
+import * as broadcaster from '../../src/broadcasting/broadcastOperations';
 
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 
-describe('broadcastFollow', function() {
+describe('broadcastReblog', function() {
   const accessToken = 'fdsgfdew';
   const username = 'jakipatryk-dev';
-  const userToFollow = 'jakipatryk';
+  const postAuthor = 'jakipatryk';
+  const postPermlink = 'i-am-jakipatryk';
 
   beforeEach(function() {
     this.broadcastOperations = sinon.stub(broadcaster, 'broadcastOperations');
   });
 
-  it('should call broadcastOperations with correct data of the user and user to follow', async function() {
+  it('should call broadcastOperations with correct data of the post to reblog', async function() {
     const operations = [
       [
         'custom_json',
@@ -22,18 +23,18 @@ describe('broadcastFollow', function() {
           required_posting_auths: [username],
           id: 'follow',
           json: JSON.stringify([
-            'follow',
+            'reblog',
             {
-              follower: username,
-              following: userToFollow,
-              what: ['blog']
+              account: username,
+              author: postAuthor,
+              permlink: postPermlink
             }
           ])
         }
       ]
     ];
 
-    await broadcastFollow(accessToken, username, userToFollow);
+    await broadcastReblog(accessToken, username, postAuthor, postPermlink);
 
     expect(this.broadcastOperations.calledWith(accessToken, operations)).to.be
       .true;
