+++ tests/broadcasting/broadcastUnfollow.spec.ts
@@ -1,20 +1,19 @@
-import { broadcastReblog } from '../src/broadcastReblog';
-import * as broadcaster from '../src/broadcastOperations';
+import { broadcastUnfollow } from '../../src/broadcasting/broadcastUnfollow';
+import * as broadcaster from '../../src/broadcasting/broadcastOperations';
 
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 
-describe('broadcastReblog', function() {
+describe('broadcastUnfollow', function() {
   const accessToken = 'fdsgfdew';
   const username = 'jakipatryk-dev';
-  const postAuthor = 'jakipatryk';
-  const postPermlink = 'i-am-jakipatryk';
+  const userToUnfollow = 'jakipatryk';
 
   beforeEach(function() {
     this.broadcastOperations = sinon.stub(broadcaster, 'broadcastOperations');
   });
 
-  it('should call broadcastOperations with correct data of the post to reblog', async function() {
+  it('should call broadcastOperations with correct data of the user and user to unfollow', async function() {
     const operations = [
       [
         'custom_json',
@@ -23,18 +22,18 @@ describe('broadcastReblog', function() {
           required_posting_auths: [username],
           id: 'follow',
           json: JSON.stringify([
-            'reblog',
+            'follow',
             {
-              account: username,
-              author: postAuthor,
-              permlink: postPermlink
+              follower: username,
+              following: userToUnfollow,
+              what: []
             }
           ])
         }
       ]
     ];
 
-    await broadcastReblog(accessToken, username, postAuthor, postPermlink);
+    await broadcastUnfollow(accessToken, username, userToUnfollow);
 
     expect(this.broadcastOperations.calledWith(accessToken, operations)).to.be
       .true;
