+++ tests/broadcasting/broadcastUpvote.spec.ts
@@ -1,10 +1,10 @@
-import { broadcastDownvote } from '../src/broadcastDownvote';
-import * as broadcaster from '../src/broadcastOperations';
+import { broadcastUpvote } from '../../src/broadcasting/broadcastUpvote';
+import * as broadcaster from '../../src/broadcasting/broadcastOperations';
 
 import { expect } from 'chai';
 import * as sinon from 'sinon';
 
-describe('broadcastDownvote', function() {
+describe('broadcastUpvote', function() {
   beforeEach(function() {
     this.broadcastOperations = sinon.stub(broadcaster, 'broadcastOperations');
   });
@@ -14,20 +14,20 @@ describe('broadcastDownvote', function() {
     const voter = 'jakipatrikko';
     const author = 'whoever';
     const permlink = 'some-permlink';
-    const weight = 500;
+    const weight = 10000;
     const operations = [
       [
         'vote',
         {
-          voter: 'jakipatrikko',
-          author: 'whoever',
-          permlink: 'some-permlink',
-          weight: -500
+          voter,
+          author,
+          permlink,
+          weight
         }
       ]
     ];
 
-    await broadcastDownvote(accessToken, voter, author, permlink, weight);
+    await broadcastUpvote(accessToken, voter, author, permlink, weight);
 
     expect(this.broadcastOperations.calledWith(accessToken, operations)).to.be
       .true;
